like/serializers.py
from rest_framework import serializers

from .models import Like


class LikeSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.email')

    class Meta:
        model = Like
        fields = ('book', 'like', 'owner',)